\name{fwi.fbp-deprecated}
\alias{fwi.fbp-deprecated}
\title{Deprecated functions in package \sQuote{fwi.fbp}}
\description{
These functions are provided for compatibility with older versions of \sQuote{fwi.fbp} only, and will be defunct at the next release.
}
\details{
  The \sQuote{fwi.fbp} package has been deprecated, and all functions have replacements within the cffdrs package.\url{https://cran.r-project.org/package=cffdrs}
  \cr
  The following functions are deprecated and will be made defunct. Use the replacements from the \sQuote{cffdrs} package indicated below (links will not work without the \sQuote{cffdrs} package installed):
  \itemize{
    \item{fwi: \code{\link[cffdrs]{fwi}}}
    \item{fwiBAT: \code{\link[cffdrs]{fwi}}}
    \item{fbp: \code{\link[cffdrs]{fbp}}}
  }
}

\author{Alan Cantin \email{Alan.Cantin@Canada.ca}}

\seealso{
  \code{\link[base]{Deprecated}}
}
